web-animations/animation-player/animation-player-cancel.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player cancel method test</title>
<meta name="assert" content="Set source to null and clears all effects associated with the previous source content">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-cancel-void">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));
    player.cancel();

    assert_equals(player.source, null, 'Value of AnimationPlayer.source attribute ' +
        'should be null');
}, 'Test AnimationPlayer.cancel() sets the AnimationPlayer.source attribute to null');

test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);

    player.currentTime = ANIMATION_END_TIME / 2;
    assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_0_5,
        'Animation effect should take place');

    player.cancel();
    assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_DEFAULT,
        'Animation effect should be cleared');
}, 'Test AnimationPlayer.cancel() clears animation effects');
</script>
</body>
